back/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.31.0

# Database
sqlalchemy==2.0.35
psycopg[binary]==3.2.3
alembic==1.13.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
bcrypt==4.2.0
cryptography==41.0.7  

# CORS
fastapi-cors==0.0.6

# Environment variables
python-dotenv==1.0.1

# Validation
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Date/Time
python-dateutil==2.9.0

# Testing (optional)
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Development tools (optional)
black==24.10.0
flake8==7.1.1
mypy==1.11.2